Set headers.status if it is unexpected for file status.
If a Partial file responds with 403, the file went missing after
initially existing. Indicate this by setting headers.status to 403 while
retaining existing headers.

To clarify, 403 from an <=Unchecked file is not included in the headers
table. A 403 being present in headers indicates that the file went
missing after initially being found, and is accompanied by the headers
from the initial response.